+++ src/aui-pagination/HISTORY.md
@@ -8,6 +8,7 @@ No registries yet.
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)
 
+* [AUI-1357](https://issues.liferay.com/browse/AUI-1357) Make aui-pagination accessible
 * [AUI-1264](https://issues.liferay.com/browse/AUI-1264) Pagination doesn't have skip to first or last page
 * [AUI-1163](https://issues.liferay.com/browse/AUI-1163) Remove unnecessary constants
 * [AUI-1106](https://issues.liferay.com/browse/AUI-1106) Add showControls attribute in a simple way